requirements.txt
pandas
numpy
psycopg2-binary
requests
python-dotenv
yfinance